V1.12
Contains a new enemy (on top of Bug's new enemy). Bullet knockback is back and a bit better. Can now set the speed of enemy bullets for individual enemy types and the 'way' that an enemy shoots/the bullets it shoots can be changed via the fireType variable. I've put some things in folders, you're bloody welcome (but it is fair it was super disorganised/hard to find anything)