MOAR optimistions
Finally made it so the player bullet only checks which enemy is closest every 6 or so physics updates rather than every single one, and it now only checks at all if you've got homing, as you'd expect it would. Started making a framework for the player having different weapons - what I'll probably do is have it so at the start of the game they choose what weapon they want to use. ATM nothing is in or working but I have started it so like get owned