Lil changes
Finally added text at the top of the screen displaying your fire delay/firerate (lower is better). Added XP back with a new levelling formula - what XP does now is increase your main stats by 3% per level (the bonus stacks additively). The purpose of it isn't to support you through the game entirely, but instead to provide you with a small little boost to your survivability. You're not really meant to notice it while playing, though playing without it and then with it makes it clear how much of a difference the little increases make. Finally, made it so you can now restart - so press R to restart when you die.