Redoing all enemy/player scripting
In order to make everything better organised, more object-oriented and more customisable, I'm transferring the code that was previously pretty much all in Player_Movement and the like into numerous scripts. This way, eventually, enemies will be able to use items (and other benefits will ensue, I'm sure).

ATM only the basic functionality for the player is done - the base shooting, taking damage, and movement is pretty much completely as it was before. Next I'm gonna make it so the enemies now use these scripts, then after that it's all about getting items working again. This is still probably gonna take a while but I think it's definitely worth doing.