stupid
ok im dumb and figured out the reason knockback was fucked 1 minute after comitting the last one (i'd been stuck on it for an hour or so last night). still no textures on the guy, but at least he has no materials because unity is designed to be as unintuitive as possible/i'm stupid.