ATGs now work with enemies
Essentially there's now a variable, owner, in DealDamage, that is given to the projectiles the player shoots that keeps track of who fired the projectile. This way the game knows where to spawn the ATG. There are two types of ATG/explosion that are otherwise identical, one that is on the playerbullets layer and the other on the enemybullets layer.