yaaaaaaaaaaayyyyyyyyyyy
ALL ITEMS (excluding stopwatch, because that would absolutely be anti-fun) now work with both the player and enemies. Currently there's no way enemies can gain items (and this will be the case for some time) but I am thinking about ways to give them items in a way that makes the most of the feature while not making it stale or frustrating. One of the current ideas is for a rare enemy type to be an otherwise normal enemy that has 5 random items. So yeah. Funkey. It's taken like 3.5 weeks to get back to a point of having the same content as the game previously did but it has definitely been worth it. I am rather pleased as you can probably tell.