see description
just committing this to say that i probably won't add any more items for a while. i have 20 more items planned, but there's a number of things i'll need to implement beforehand for some of them, and in all honestly i'm bored of adding items lmao. i just finished off the contact item because it had been sitting in ITEMLIST for ages but hadn't been implemented. also possible i'll be taking a bit of a break from vampire pooers for a bit since i've spent a lot of time on it in the past couple of weeks.