Poisson
Big updates to the status effect system. Now, ALL status effects are handled within a single script (called Statuses), and the icon showing the status effect is a single object that chooses if it represents poison or bleed. The positioning of the icon and text depends on how many status effects the player or enemy has. The order at which the effects are displayed (i.e. which one is displayed to the left, which one to the right) is determined by which one was applied first.

Oh yeah, and added a poison upgrade (when an enemy dies, creates an explosion that poisons all nearby enemies). To do this, there's also now a delegate on the master object that tracks all on-death effects. Later I should make an additional event manager but for enemy effects only, because atm enemies can use the poison explosion afaik.

This took fucking ages to get working properly, but I am quite pleased with how it turned out/how it works.