icon shite
Added an icon for when enemies are slowed (by wapant). Icon positioning was a bit fucked so that's fixed as well. Status icons now rendered in front of spinner, but note I really should just pivot to using sorting layers in future rather than fucking around with the z-coordinate of things.