molar
Mole still broke a bit if you killed the one that started the lazer before it got to fire, so that's now fixed. Might still be one way to fuck them up but I think they work well enough atm that I'm happy to move on from em for now.