healing stuff
Added a new script for healing enemies/player (forgot to give it to enemies atm tho lmao), anything that heals should just call the heal function on the player with the amount to heal with. This is done so items that increase/decrease your healing ability can exist. Slightly changed poison colour text and added healing text.