Finished items and CURSED items
Finally added the sprites for the new items. Added new cursed items, that add 2 of the item to you but also give one of the item to ALL enemies. They have a 1/20 chance to appear and are shown as a bigger version of the normal item (not final appearance, obviously). Need to add tags for non-enemy items, so cursed items can't appear as non-enemy items.