Lil fixers
So, for items that have a kinda complex start (hence running start on every new bullet would be slow) there's now a check for if the new bullet is a clone of the prefab or the prefab itself, this bool is kept in dealdamge and set by attack. This does two things - the aforementioned (very slight) optimisation, and DMG ups and soy milk etc. now do their shit the proper number of times. Also fixed dagger throw with the current bullet system and made it scale differently since the previous way wasn't super useful.