Number-related things
There is now a new world-space canvas, used for shit like the damage numbers and showing stacks of DOTs. Simpler to program for and doesn't have fucked up scaling with different resolutions. Also, damage resistances/types are now implemented, so things like creep and poison now deal poison damage, bleed has its own type, there's explosion damage, and so on. The type of a damage source is stored in dealdamage, but unless it's just normal (physical) damage, the type is set by a script - i.e. the explosion script sets its damage type, the creep script sets its type, and so on. The Hurty method takes the type and applies the resistance.