De-nightmared item adding
Reworked everything involving giving items, so now it works generally. There are a couple new things that need to be set in ItemDescriptions when you add a new item (one relating to what gets the item, i.e. does it get attached to bullets and/or beings or the master object and the like, and another relating to whether a new instance of the script should be added when stacking the item or not). For items that stack via increasing the instances variable, there's a new method (it's in the word doc) that you have to put in (doesn't require editing cos it's general). There are a couple other general fixes and improvements to items, how they're added, and so on, and as far as I know it all works.