Okie
Orbitals are properly fixed for enemies now. Alt dodge abilities are being removed, because they are not that versatile and the better ideas I had for that work better as the kinda Dead Cells-esque items I have planned. Following on from that, banned items is no longer a thing, since all items can stack indefinitely anyway now, and because rarely being offered the weapon you already have is not a big deal (also just complicates things in co-op). Also also splitting infinitely with melee is removed, it was far too stupid lol.