Simplification, baby
So because they were done ages ago, the code inside of the oncollisonenter and ontriggerstay in HPDamageDie were kinda stupid and had repeated code. So, now what happens is the same in both the trigger and collision methods (and is way shorter, just setting ignorecollision and then calling a new method in DealDamage). This new method is called on the object responsible for the hit, and essentially takes the base damage and applies any damage modifiers to it (damage modifiers are from damage ups, soy, explosion damage falloff, etc.), then calls Hurty on the victim with the respective damage amount. This massively simplifies the damage ups which were previously nightmarishly coded. There may be a couple of damage sources that are now a little fucked and do the wrong amount of damage, but I'll fix those when they appear. Some other fixes were done like player orbitals double firing (I fucking swear I have to fix the orbitals all the damn time, I don't even think they're poorly coded I just think I keep doing a goof lmao) and the player is no longer slow upon taking the bat.