UI Mostly-ish-kinda done
Added minimap and wave progress UI. For minimap, rocks and TNT have a spriterenderer child on a layer only the minimap camera renders, which it uses to create the map layout. To make POIs appear on the map when the player finds them, add a sprite (which will appear on the map) as a child of the POI, with layer set to the correct one and tag pickupforplayer. On trigger enter with another pickupforplayer, the spriterenderer is enabled, i.e. the sprite appears on the map. this is done with the humphrey/buggo icon.