roomz systemme
so the rooms system is in. there are SO's for each room layout, and one 'master' SO for the room layouts for each level. the master SO is to be added to the level generator by hand, which then picks the right room set at runtime and does a cumsum to pick a room. goes hard, screenshot if you want.