Meat
Just made the game feel 100% better to play by introducing more new definable parameters for abilities and re-doing screenshake. New parameters are the amount of screenpush to apply, and a bool for whether to apply the screenpush in the direction opposite to the direction of firing or do it in a random direction.